Fix issue #45 Crash when closing after having opened scatterplot plugin
The crash was caused by having multiple destructor calls on one an the same instance of various QWidget derived class types.

The destructor of these objects was called _both_ via Qt's ownership logic (typically because they were passed as argument to `QLayout::addWidget`), _and_ by the destructor of the objects that "had" them stored as data member.

The latter is avoided by this commit, by creating those objects by `new`, and storing references (`&`) to the objects as data members, instead of storing the objects themselves as data members.