PointArrayObject "set" member functions avoid setting "dirty" true
The `PointArrayObject` member functions `setPositions`, `setHighlights`, `setScalars`, and `setColors` now avoid setting "dirty" true when the data of the `std::vector` passed as argument is bitwise equal to the data of the corresponding member variable.

They do however, set "dirty" true when the argument is an empty `std::vector`, and the corresponding member variable is non-empty. In this case, the member variable is also effectively set to an empty vector. This behavior is different than before.

Discussed with Julian Thijssen.