Add Points::visitSourceData, implemented by custom index functions
`Points::visitSourceData` allows accessing the source data. When either the source data or the derived data (or both) are a subset, `visitSourceData` specifically allows access to that subset.

Internally it is implemented by specifying a custom index function for the range of `PointDataIterator` iterators. The index function specifies how to get the index into the internal raw data buffer.